requirements.txt
phidata 
nest_asyncio 
beautifulsoup4 
openai 
duckduckgo-search 
sqlalchemy 
pgvector 

pypdf 
ollama

streamlit
libpq